Create a VC project for the test binary.
Also add a getopt() implementation and a unistd.h from stackoverflow
to allow the existing code in test.c to build. (The normal unistd.h and
getopt() should still be used on systems that provide it. VC does not, so
that's why it's added.)

Also add a version.bat batch file that creates a VERSION.H file for VC
to include in the build. (Defines VERSION in the preprocessor, Batch file
pulls the version number from version.mk)