meta/ogg_vorbis: Support LOOPMS tags
Teach the Ogg Vorbis parser to interpret the value from LOOPMS tags as a
loop start position, measured in milliseconds (instead of the typical
samples). Converting this value to samples requires the sample rate, so
we must now read the sample rate before interpreting the tags.

These tags are used in Sonic Robo Blast 2 (as of v2.2.6).