fix some ongakukan .adp
at least one adp file from midousuji-sen was found to have a "fact" chunk at offset 0x24 instead of the usual 0x26. this caused vgmstream to not play said file at all.

not only does this commit fixes this, it also adds additional checks to the "adp_ongakukan" meta.

this includes the introduction of the "diff" variable whose purpose is to subtract the difference between expected_size and pcm_size. meta will now only proceed if diff is between 0 and 2 (sorta).

as for the "fix" itself, it required the introduction of the "fact_offset" variable.

fact_offset adds both fmt_offset and fmt_size so that "fact" chunk signature can be read (and checked against adp file within the meta) properly instead of always requiring "fact" to be at 0x26.

fact_offset is also checked if it goes below 0x24 and above 0x28.