xsb: Handle endianness for variation count and flags
I was having all sorts of crashes when trying to use vgmstream to play
audio from .xwb/.wsb files from an old PC game. When I enabled logs I
found negative stream indices -- an indication of the file being parsed
erroneously.

After a lot of debugging and comparing vgmstream's parser with wine's, I
realized the variation count and flags fields are treated as a 32-bit
bitstruct in wine.

The current code in vgmstream for parsing these fields has a comment
about MonoGame reading them in the opposite order but it arguably being
wrong since vgmstream's order works for an Xbox 360 game.

Notice that the Xbox 360 is a big endian platform whereas x86 is a
little endian platform. Reading both fields as a 32-bit bitstruct
results in the same result as before for big endian and fixes the
decoding of at least one little endian file.